Basic funcionality implemented
For now you can search movies by selecting a month.
The movies will be stored as a JSON object in the movies variable.
For now the object is just there, there is no visualization of the
search results.
The .php file is necessary because you cannot get data from external
servers in JavaScript